[branch-4.14] Fix the problem that the abnormal file causes the bookie GC fail (#3611)
Descriptions of the changes in this PR:

Motivation
When bookie GC encounters an entryLog file with disordered data, a IllegalArgumentException are not caught, resulting in the death of the GC thread and the end of the GC process.
This problem is encountered again in the next GC, resulting in the subsequent entrylog being unable to GC.
The problem described in this issue(https://github.com/apache/bookkeeper/issues/3604) will cause the entryLog file to be disordered, and will cause the IllegalArgumentException when parsing the entrylog file.

Master Issue:
https://github.com/apache/bookkeeper/issues/3607 : Describes the problem to be solved by this pr.
https://github.com/apache/bookkeeper/issues/3604 : Describes the phenomenon and causes of disordered entrylog files.

(cherry picked from commit 7460cf5ad25c7c6e7b3b77045c7992044e3db2f8)